Move updateCheck function to `update-check` package
