+++ api/package.json
@@ -80,7 +80,6 @@
 		"@directus/storage-driver-gcs": "workspace:*",
 		"@directus/storage-driver-local": "workspace:*",
 		"@directus/storage-driver-s3": "workspace:*",
-		"@directus/update-check": "workspace:*",
 		"@directus/utils": "workspace:*",
 		"@godaddy/terminus": "4.12.0",
 		"@rollup/plugin-alias": "5.0.0",
