+++ directus/package.json
@@ -50,26 +50,11 @@
 		"./package.json": "./package.json"
 	},
 	"bin": {
-		"directus": "dist/cli.js"
-	},
-	"scripts": {
-		"build": "tsc --project tsconfig.json",
-		"dev": "tsc --watch"
+		"directus": "cli.js"
 	},
 	"dependencies": {
 		"@directus/api": "workspace:*",
-		"boxen": "7.1.0",
-		"chalk": "5.2.0",
-		"find-cache-dir": "4.0.0",
-		"node-fetch-cache": "3.1.3"
-	},
-	"devDependencies": {
-		"@directus/tsconfig": "workspace:*",
-		"@npm/types": "1.0.2",
-		"@types/find-cache-dir": "3.2.1",
-		"@types/node-fetch-cache": "3.0.0",
-		"tsup": "6.7.0",
-		"typescript": "5.0.4"
+		"@directus/update-check": "workspace:*"
 	},
 	"engines": {
 		"node": ">=18.0.0"
