--- directus/src/cli.ts
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-import { updateNotifier } from './update-notifier.js';
-
-await updateNotifier();
-import('@directus/api/cli/run.js');
