+++ packages/update-check/src/index.ts
@@ -1,37 +1,80 @@
-import { URL } from 'node:url';
-import { gte } from 'semver';
-import { fetch } from 'undici';
+import boxen, { type Options as BoxenOptions } from 'boxen';
+import chalk from 'chalk';
+import type { Manifest } from '@npm/types';
+import findCacheDirectory from 'find-cache-dir';
+import { fetchBuilder, FileSystemCache } from 'node-fetch-cache';
 
-/**
- * Check if a given package version is the most up to date release of that package. Returns the
- * latest version string if package is not up to date, returns null if package is up to date.
- */
-export const isUpToDate = async (name: string, version: string): Promise<string | null> => {
-	const url = new URL(name, 'https://registry.npmjs.org');
+const cacheDirectory = findCacheDirectory({ name: 'directus' });
 
-	const response = await fetch(url, {
-		headers: {
-			accept: 'application/vnd.npm.install-v1+json; q=1.0, application/json; q=0.8, */*',
-		},
-	});
+const fetch = fetchBuilder.withCache(
+	new FileSystemCache({ ttl: 60 * 60 * 60, ...(cacheDirectory && { cacheDirectory }) })
+);
 
-	if (!response.ok) {
-		throw new Error(`Couldn't find latest version for package "${name}": ${response.status} ${response.statusText}`);
+export async function updateCheck(currentVersion: string) {
+	let packageManifest: Manifest | undefined = undefined;
+
+	try {
+		const response = await fetch('https://registry.npmjs.org/directus', {
+			headers: { accept: 'application/vnd.npm.install-v1+json; q=1.0, application/json; q=0.8, */*' },
+		});
+
+		if (!response.ok) {
+			response.ejectFromCache();
+			return;
+		}
+
+		packageManifest = await response.json();
+	} catch (error) {
+		// Ignore
 	}
 
-	const packageInformation = (await response.json()) as { 'dist-tags': { [key: string]: string } };
-
-	const latest = packageInformation?.['dist-tags']?.['latest'];
-
-	if (!latest) {
-		throw new Error(`Couldn't find latest version for package "${name}"`);
+	if (!packageManifest) {
+		return;
 	}
 
-	const upToDate = gte(version, latest);
+	const latestVersion = packageManifest['dist-tags']['latest'];
 
-	if (upToDate === false) {
-		return latest;
+	if (!latestVersion || currentVersion === latestVersion) {
+		return;
 	}
 
-	return null;
-};
+	const allVersions = Object.keys(packageManifest.versions);
+	const indexOfCurrent = allVersions.indexOf(currentVersion);
+	const indexOfLatest = allVersions.indexOf(latestVersion);
+
+	const versionDifference =
+		indexOfCurrent !== -1 && indexOfLatest !== -1 ? Math.abs(indexOfLatest - indexOfCurrent) : null;
+
+	const message = [
+		chalk.bold(`Update available!`),
+		'',
+		chalk.bold(`${chalk.red(currentVersion)} → ${chalk.green(latestVersion)}`),
+		...(versionDifference
+			? [chalk.dim(`${versionDifference} ${versionDifference > 1 ? 'versions' : 'version'} behind`)]
+			: []),
+		'',
+		'More information:',
+		chalk.blue(`https://github.com/directus/directus/releases`),
+	];
+
+	let borderColor;
+
+	if (versionDifference && versionDifference > 5) {
+		borderColor = 'red';
+	} else if (versionDifference && versionDifference > 2) {
+		borderColor = 'yellow';
+	} else {
+		borderColor = 'magenta';
+	}
+
+	const boxenOptions: BoxenOptions = {
+		padding: 1,
+		margin: 1,
+		align: 'center',
+		borderColor,
+		borderStyle: 'round',
+	};
+
+	// eslint-disable-next-line no-console
+	console.warn(boxen(message.join('\n'), boxenOptions));
+}
